requirements.txt
setuptools~=75.1.0
ultralytics~=8.2.97
torch~=2.4.1
matplotlib~=3.9.2
opencv-python~=4.10.0.84
numpy~=1.26.4
torchvision~=0.19.1
lime~=0.2.0.1
pillow~=10.4.0
scikit-image~=0.24.0
logging~=0.4.9.6
comet_ml~=3.45.1
ttach~=0.0.3
shap~=0.46.0
onnxruntime~=1.19.2
onnx~=1.16.2